// packages/shared/src/server/types/status.d.ts
export declare enum ProcessingStatus {
    IN_QUEUE = "in_queue",
    FETCHING_TRANSCRIPT = "fetching_transcript",
    GENERATING_SUMMARY = "generating_summary",
    COMPLETED = "completed",
    FAILED = "failed"
}
export type ClientStatus = 'IN_QUEUE' | 'FETCHING_TRANSCRIPT' | 'GENERATING_SUMMARY' | 'COMPLETED' | 'FAILED';
//# sourceMappingURL=status.d.ts.map